Enable NPC loot inventory to be sorted on init.
For some reason, vanilla UIInventory.LMode_Init function has an exception for NPC loot boxes (corpses) that prevents inventory reset on startup. Skipping or executing this reset seems to have no effect, but it prevents SortingPlus to do its thing when the loot screen comes up. If you want the NPC loot to be sorted, you either have to refresh the inventory (F5) or close and reopen the loot screen.

Well, not anymore. Removing the (seemingly unnecessary) check for npc_box enables SortingPlus to sort loot views like the others so npc loot is not all over the place when first seen.